Build tree up-front so it's in treeNodesByLine
This fixes the issue where we couldn't expand/highlight something in the tree because when the parent was's parent was requested we were cleaing out the treeNodesByLine and then only re-building it as far as we'd walked down the tree (which is we were collapsed, was not the full tree).

This builds the whole tree up-front which is more expensive, but since we have the tree expanded by default may work out better because we're not constantly rebuilding the children and the treeNodesByLine collection.